Change ACE Hint preview
